Log hex dumps of mc -> ws packets as well
